requirements.txt
aiohttp>=3.9.0
azure-identity>=1.16.0
